implement managed.py using abstraction
